+++ src/components/show_bundle/bundle_details.tsx
